HT-402: Added unit tests for AudioRecorder (and refactored/moved existing unit tests for other Media classes)
Fixed some problems and improved/added comments based on code review and a flaw detected by unit test.
