+++ SIL.Media.Tests/AudioRecorderTests.cs
@@ -1,493 +1,178 @@
-using System;
-using System.ComponentModel;
-using System.Diagnostics;
 using System.IO;
-using System.Media;
+using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
 using NUnit.Framework;
 using SIL.IO;
-using SIL.Media.Tests.Properties;
-using SIL.TestUtilities;
+using SIL.Media.Naudio;
 
 namespace SIL.Media.Tests
 {
 	// Some of these tests require a speaker. Others require a microphone.
 	// None of them will work if neither a speaker nor a microphone is available.
 	[TestFixture]
-	[NUnit.Framework.Category("SkipOnTeamCity")]
-	[NUnit.Framework.Category("AudioTests")]
+	[Category("SkipOnTeamCity")]
+	[Category("AudioTests")]
 	public class AudioRecorderTests
 	{
+		
 		[Test]
-		public void Construct_FileDoesNotExist_OK()
-		{
-			// ReSharper disable once UnusedVariable
-			using (var x = AudioFactory.CreateAudioSession(Path.GetRandomFileName())) { }
-		}
-
-		[Test]
-		public void Construct_FileDoesExistButEmpty_OK()
+		public void BeginRecording_OnNonUiThread_Throws()
 		{
 			using (var f = new TempFile())
 			{
-				// ReSharper disable once UnusedVariable
-				using (var x = AudioFactory.CreateAudioSession(f.Path)) { }
-			}
-		}
-
-
-		[Test]
-		public void Construct_FileDoesNotExist_DoesNotCreateFile()
-		{
-			var path = Path.GetRandomFileName();
-			// ReSharper disable once UnusedVariable
-			using (var x = AudioFactory.CreateAudioSession(path)) { }
-			Assert.IsFalse(File.Exists(path)); // File doesn't exist after disposal
-		}
-
-		[Test]
-		public void StopRecording_NotRecording_Throws()
-		{
-			using (var x = AudioFactory.CreateAudioSession(Path.GetRandomFileName()))
-			{
-				Assert.Throws<ApplicationException>(() => x.StopRecordingAndSaveAsWav());
-			}
-		}
-
-		[Test]
-		public void StopPlaying_WhilePlaying_Ok()
-		{
-			using (var session = new RecordingSession(1000))
-			{
-				session.Recorder.Play();
-				Thread.Sleep(100);
-				session.Recorder.StopPlaying();
-			}
-		}
-
-		[Test]
-		public void Record_AfterRecordThenStop_Ok()
-		{
-			using (var session = new RecordingSession(100))
-			{
-				session.Recorder.StartRecording();
-				Thread.Sleep(100);
-				session.Recorder.StopRecordingAndSaveAsWav();
-			}
-		}
-
-		[Test]
-		public void Play_WhileRecording_Throws()
-		{
-			using (var session = new RecordingSession())
-			{
-				Assert.Throws<ApplicationException>(() => session.Recorder.Play());
-			}
-		}
-
-		[Test]
-		public void CanRecord_FileDoesNotExist_True()
-		{
-			using (var x = AudioFactory.CreateAudioSession(Path.GetRandomFileName()))
-			{
-				Assert.IsTrue(x.CanRecord);
-			}
-		}
-
-		[Test]
-		public void CanStop_NonExistentFile_False()
-		{
-			using (var x = AudioFactory.CreateAudioSession(Path.GetRandomFileName()))
-			{
-				Assert.IsFalse(x.CanStop);
-			}
-		}
-
-		[Test]
-		public void CanRecord_ConstructWithEmptyFile_True()
-		{
-			using (var f = new TempFile())
-			{
-				using (var x = AudioFactory.CreateAudioSession(f.Path))
+				using (var recorder = new AudioRecorder(1))
 				{
-					Assert.IsTrue(x.CanRecord);
+					recorder.SelectedDevice = RecordingDevice.Devices.First() as RecordingDevice;
+					Assert.That(() => recorder.BeginRecording(f.Path, false), Throws.Exception);
 				}
 			}
 		}
 
 		[Test]
-		public void Play_FileEmpty_Throws()
+		public void BeginRecording_ThenStop_RecordingSavedToFile()
 		{
-			using (var f = new TempFile())
+			using (var ctrl = new Control())
 			{
-				using (var x = AudioFactory.CreateAudioSession(f.Path))
-				{
-					Assert.Throws<FileLoadException>(() => x.Play());
-				}
-			}
-		}
-
-		[Test]
-		public void Play_FileDoesNotExist_Throws()
-		{
-			using (var x = AudioFactory.CreateAudioSession(Path.GetRandomFileName()))
-			{
-				Assert.Throws<FileNotFoundException>(() => x.Play());
-			}
-		}
-
-		[Test]
-		public void CanStop_WhilePlaying_True()
-		{
-			using (var session = new RecordingSession(1000))
-			{
-				session.Recorder.Play();
-				Thread.Sleep(100);
-				Assert.IsTrue(session.Recorder.CanStop);
-			}
-		}
-
-
-		[Test]
-		public void RecordAndStop_FileAlreadyExists_FileReplaced()
-		{
-			using (var f = new TempFile())
-			{
-				var oldInfo = new FileInfo(f.Path);
-				var oldLength = oldInfo.Length;
-				Assert.AreEqual(0, oldLength);
-				var oldTimestamp = oldInfo.LastWriteTimeUtc;
-				using (var x = AudioFactory.CreateAudioSession(f.Path))
-				{
-					x.StartRecording();
-					Thread.Sleep(1000);
-					x.StopRecordingAndSaveAsWav();
-				}
-				var newInfo = new FileInfo(f.Path);
-				Assert.Greater(newInfo.LastWriteTimeUtc, oldTimestamp);
-				Assert.Greater(newInfo.Length, oldLength);
-			}
-		}
-
-		[Test]
-		public void IsRecording_WhileRecording_True()
-		{
-			using (var f = new TempFile())
-			{
-				using (var x = AudioFactory.CreateAudioSession(f.Path))
-				{
-					x.StartRecording();
-					Thread.Sleep(100);
-					Assert.IsTrue(x.IsRecording);
-					x.StopRecordingAndSaveAsWav();
-				}
-			}
-		}
-
-		[Test]
-		[Platform(Exclude = "Linux", Reason = "AudioAlsaSession doesn't implement ISimpleAudioWithEvents")]
-		public void RecordThenPlay_SmokeTest()
-		{
-			using (var f = new TempFile())
-			{
-				var w = new BackgroundWorker();
-				// ReSharper disable once RedundantDelegateCreation
-				w.DoWork += new DoWorkEventHandler((o, args) => SystemSounds.Exclamation.Play());
-
-				using (var x = AudioFactory.CreateAudioSession(f.Path))
-				{
-					x.StartRecording();
-					w.RunWorkerAsync();
-					Thread.Sleep(1000);
-					x.StopRecordingAndSaveAsWav();
-					bool stopped = false;
-					bool isPlayingInEventHandler = false;
-					(x as ISimpleAudioWithEvents).PlaybackStopped += (o, args) =>
-					{
-						// assert here is swallowed, probably because not receiving exceptions from background worker.
-						// We want to check that isPlaying is false even during the event handler.
-						isPlayingInEventHandler = x.IsPlaying;
-						stopped = true;
-					};
-					var watch = Stopwatch.StartNew();
-					x.Play();
-					while (!stopped)
-					{
-						Thread.Sleep(20);
-						Application.DoEvents();
-						if (watch.ElapsedMilliseconds > 2000)
-						{
-							x.StopPlaying();
-							Assert.Fail("stop event not received");
-						}
-					}
-					// After playback is stopped we shouldn't be reporting that it is playing
-					Assert.That(isPlayingInEventHandler, Is.False);
-					Assert.That(x.IsPlaying, Is.False);
-				}
-			}
-		}
-
-		[Test]
-		public void Play_GiveThaiFileName_ShouldHearTwoSounds()
-		{
-			using (var d = new TemporaryFolder("palaso media test"))
-			{
-				var soundPath = d.Combine("ก.wav");
-				File.Create(soundPath).Close();
-				using (var f = TempFile.TrackExisting(soundPath))
-				{
-					var w = new BackgroundWorker();
-					// ReSharper disable once RedundantDelegateCreation
-					w.DoWork += new DoWorkEventHandler((o, args) => SystemSounds.Exclamation.Play());
-
-					using (var x = AudioFactory.CreateAudioSession(f.Path))
-					{
-						x.StartRecording();
-						w.RunWorkerAsync();
-						Thread.Sleep(2000);
-						x.StopRecordingAndSaveAsWav();
-					}
-
-					using (var y = AudioFactory.CreateAudioSession(f.Path))
-					{
-						y.Play();
-						Thread.Sleep(1000);
-						y.StopPlaying();
-					}
-				}
-			}
-		}
-
-
-		/// <summary>
-		/// for testing things while recording is happening
-		/// </summary>
-		class RecordingSession : IDisposable
-		{
-			private TempFile _tempFile;
-			private ISimpleAudioSession _recorder;
-
-			public RecordingSession()
-			{
-				_tempFile = new TempFile();
-				_recorder = AudioFactory.CreateAudioSession(_tempFile.Path);
-				_recorder.StartRecording();
-				Thread.Sleep(100);
-			}
-
-			public RecordingSession(int millisecondsToRecordBeforeStopping)
-				: this()
-			{
-				if (millisecondsToRecordBeforeStopping == 0) millisecondsToRecordBeforeStopping = 1000;
-				Thread.Sleep(millisecondsToRecordBeforeStopping);
-				_recorder.StopRecordingAndSaveAsWav();
-			}
-
-			public ISimpleAudioSession Recorder
-			{
-				get { return _recorder; }
-			}
-
-			public void Dispose()
-			{
+				bool recordingStopped = false;
+				var f = new TempFile();
 				try
 				{
-					if (_recorder.IsRecording)
+					var recorder = new AudioRecorder(1);
+					// Note: BeginRecording
+					ctrl.HandleCreated += delegate
 					{
-						_recorder.StopRecordingAndSaveAsWav();
+						recorder.SelectedDevice =
+							RecordingDevice.Devices.First() as RecordingDevice;
+						recorder.BeginRecording(f.Path, false);
+						Thread.Sleep(100);
+						recorder.Stop();
+					};
+
+					recorder.Stopped += delegate
+					{
+						Assert.That(f.Path, Does.Exist);
+						Assert.That(new FileInfo(f.Path).Length, Is.GreaterThan(0));
+						recordingStopped = true;
+						recorder.Dispose();
+					};
+
+					ctrl.CreateControl();
+
+					for (var i = 0; i < 3 && !recordingStopped; i++)
+					{
+						Application.DoEvents();
+						Thread.Sleep(200);
 					}
+
+					Assert.That(recordingStopped);
 				}
 				finally
 				{
-					_recorder.Dispose();
-					_tempFile.Dispose();
+					f.Dispose();
 				}
 			}
 		}
 
 		[Test]
-		public void CanStop_WhileRecording_True()
+		public void BeginRecording_WhileRecording_ThrowsInvalidOperationException()
 		{
-			using (var session = new RecordingSession())
+			using (var ctrl = new Control())
 			{
-				Assert.IsTrue(session.Recorder.CanStop);
-			}
-		}
-
-		[Test]
-		public void CanPlay_WhileRecording_False()
-		{
-			using (var session = new RecordingSession())
-			{
-				Assert.IsFalse(session.Recorder.CanPlay);
-			}
-		}
-
-		[Test]
-		public void CanRecord_WhileRecording_False()
-		{
-			using (var session = new RecordingSession())
-			{
-				Assert.IsFalse(session.Recorder.CanRecord);
-			}
-		}
-
-		[Test]
-		public void StartRecording_WhileRecording_Throws()
-		{
-			using (var session = new RecordingSession())
-			{
-				Assert.Throws<ApplicationException>(() => session.Recorder.StartRecording());
-			}
-		}
-
-		[Test]
-		public void CanRecord_WhilePlaying_False()
-		{
-			using (var session = new RecordingSession(1000))
-			{
-				session.Recorder.Play();
-				Thread.Sleep(100);
-				Assert.IsTrue(session.Recorder.IsPlaying);
-				Assert.IsFalse(session.Recorder.CanRecord);
-			}
-		}
-
-		[Test]
-		public void CanPlay_WhilePlaying_False()
-		{
-			using (var session = new RecordingSession(1000))
-			{
-				session.Recorder.Play();
-				Thread.Sleep(100);
-				Assert.IsFalse(session.Recorder.CanPlay);
-			}
-		}
-
-		[Test]
-		public void IsRecording_AfterRecording_False()
-		{
-			using (var f = new TempFile())
-			{
-				using (ISimpleAudioSession x = RecordSomething(f))
+				bool recordingStopped = false;
+				var f = new TempFile();
+				var recorder = new AudioRecorder(1);
+				try
 				{
-					Assert.IsFalse(x.IsRecording);
+					// Note: BeginRecording
+					ctrl.HandleCreated += delegate
+					{
+						recorder.SelectedDevice =
+							RecordingDevice.Devices.First() as RecordingDevice;
+						recorder.BeginRecording(f.Path, false);
+					};
+
+					ctrl.CreateControl();
+
+					Thread.Sleep(10);
+					Assert.That(() => recorder.BeginRecording("blah.wav"), Throws.InvalidOperationException);
+
+				}
+				finally
+				{
+					recorder.Dispose();
+					f.Dispose();
 				}
 			}
 		}
 
 		[Test]
-		public void RecordThenStop_CanPlay_IsTrue()
+		public void BeginMonitoring_WhileRecording_ThrowsInvalidOperationException()
 		{
-			using (var f = new TempFile())
+			using (var ctrl = new Control())
 			{
-				using (ISimpleAudioSession x = RecordSomething(f))
+				bool recordingStopped = false;
+				var f = new TempFile();
+				var recorder = new AudioRecorder(1);
+				try
 				{
-					Assert.IsTrue(x.CanPlay);
+					// Note: BeginRecording
+					ctrl.HandleCreated += delegate
+					{
+						recorder.SelectedDevice =
+							RecordingDevice.Devices.First() as RecordingDevice;
+						recorder.BeginRecording(f.Path, false);
+					};
+
+					ctrl.CreateControl();
+
+					Thread.Sleep(10);
+					Assert.That(() => recorder.BeginMonitoring(), Throws.InvalidOperationException);
+
+				}
+				finally
+				{
+					recorder.Dispose();
+					f.Dispose();
 				}
 			}
 		}
 
 		[Test]
-		public void RecordThenPlay_OK()
+		public void BeginRecording_OnNonUiThreadAfterMonitoringStartedOnUIThread_NoException()
 		{
-			using (var f = new TempFile())
+			using (var ctrl = new Control())
 			{
-				using (ISimpleAudioSession x = RecordSomething(f))
+				bool monitoringStarted = false;
+				var f = new TempFile();
+				var recorder = new AudioRecorder(1);
+				try
 				{
-					x.Play();
+					ctrl.HandleCreated += delegate
+					{
+						recorder.SelectedDevice =
+							RecordingDevice.Devices.First() as RecordingDevice;
+						recorder.BeginMonitoring();
+						monitoringStarted = true;
+					};
+
+					ctrl.CreateControl();
+
+					for (var i = 0; i < 3 && !monitoringStarted; i++)
+					{
+						Application.DoEvents();
+						Thread.Sleep(200);
+					}
+
+					recorder.BeginRecording(f.Path);
 					Thread.Sleep(100);
-					x.StopPlaying();
-				}
-			}
-		}
-
-		private ISimpleAudioSession RecordSomething(TempFile f)
-		{
-			var x = AudioFactory.CreateAudioSession(f.Path);
-			x.StartRecording();
-			Thread.Sleep(100);
-			x.StopRecordingAndSaveAsWav();
-			return x;
-		}
-
-		[Test]
-		public void Play_DoesPlay ()
-		{
-			using (var file = TempFile.FromResource(Resources.finished, ".wav"))
-			{
-				using (var x = AudioFactory.CreateAudioSession(file.Path))
-				{
-					Assert.DoesNotThrow(() => x.Play());
-					Assert.DoesNotThrow(() => x.StopPlaying());
-				}
-			}
-		}
-
-		[Test]
-		[Platform(Exclude = "Linux", Reason = "AudioAlsaSession doesn't implement ISimpleAudioWithEvents")]
-		public void Play_DoesPlayMp3_SmokeTest()
-		{
-			// file disposed after playback stopped
-			var file = TempFile.FromResource(Resources.ShortMp3, ".mp3");
-			using (var x = AudioFactory.CreateAudioSession(file.Path))
-			{
-				(x as ISimpleAudioWithEvents).PlaybackStopped += (e, f) =>
-				{
-					Debug.WriteLine(f);
+					recorder.Stop();
 					Thread.Sleep(100);
-					file.Dispose();
-				};
-				Assert.That(x.IsPlaying, Is.False);
-				Assert.DoesNotThrow(() => x.Play());
-				Assert.That(x.IsPlaying, Is.True);
-				Assert.DoesNotThrow(() => x.StopPlaying());
-				Assert.That(x.IsPlaying, Is.False);
-			}
-		}
-
-		[Test]
-		public void Record_DoesRecord ()
-		{
-			using (var folder = new TemporaryFolder("Record_DoesRecord"))
-			{
-				string fpath = Path.Combine(folder.Path, "dump.ogg");
-				using (var x = AudioFactory.CreateAudioSession(fpath))
-				{
-					Assert.DoesNotThrow(() => x.StartRecording());
-					Assert.IsTrue(x.IsRecording);
-					Thread.Sleep(1000);
-					Assert.DoesNotThrow(() => x.StopRecordingAndSaveAsWav());
 				}
-			}
-		}
-
-		[Test]
-		[NUnit.Framework.Category("ByHand")]
-		[Explicit] // This test is to be run manually to test long recordings. After the first beep, recite John 3:16"
-		public void Record_LongRecording()
-		{
-			using (var folder = new TemporaryFolder("Record_LongRecording"))
-			{
-				string fpath = Path.Combine(folder.Path, "long.wav");
-				using (var x = AudioFactory.CreateAudioSession(fpath))
+				finally
 				{
-					SystemSounds.Beep.Play();
-					Assert.DoesNotThrow(() => x.StartRecording());
-					Assert.IsTrue(x.IsRecording);
-					Thread.Sleep(10000); // Records 10 seconds
-					Assert.DoesNotThrow(() => x.StopRecordingAndSaveAsWav());
-					SystemSounds.Beep.Play();
-					Assert.IsTrue(x.CanPlay);
-					Assert.DoesNotThrow(() => x.Play());
-					Thread.Sleep(4000); // Plays the first 4 seconds
-					Assert.DoesNotThrow(() => x.StopPlaying());
-					Thread.Sleep(500); // Pause
-					Assert.DoesNotThrow(() => x.Play());
-					Thread.Sleep(6000); // Plays the first 6 seconds
-					Assert.DoesNotThrow(() => x.StopPlaying());
+					recorder.Dispose();
+					f.Dispose();
 				}
 			}
 		}
