+++ SIL.Media/AudioFactory.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Diagnostics;
-using System.Globalization;
 using System.IO;
 using System.Reflection;
 using SIL.Media.AlsaAudio;
